+++ src/main/proximitymine_reset.c
@@ -40,7 +40,7 @@ void proximitymine_resetToIdle(ProximityMineObject* obj)
         f32 dist = state->triggerDistance - lbl_803E6774;
         spawnExplosion(obj, dist * lbl_803DC24C + lbl_803E6770, 1, 1, 0, 1, 0, 1, 0);
     }
-    ObjHitbox_SetCapsuleBounds(obj, (s32)state->triggerDistance, -5, 10);
+    ObjHitbox_SetCapsuleBounds(obj, state->triggerDistance, -5, 10);
     ObjHits_SetHitVolumeSlot((u32)obj, 13, 1, 0);
     ObjHits_EnableObject((u32)obj);
     if (state->effectHandle != NULL)
